test/identity-test/python/requirements.txt
azure-identity
azure-mgmt-resource
